+++ Geometry/MuonCommonData/plugins/DDGEMAngular.cc
@@ -5,13 +5,47 @@
 
 #include <cmath>
 #include <algorithm>
-#include <iostream>
+#include <map>
+#include <string>
+#include <vector>
 
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/PluginManager/interface/PluginFactory.h"
 #include "DetectorDescription/Core/interface/DDLogicalPart.h"
 #include "DetectorDescription/Core/interface/DDCurrentNamespace.h"
 #include "DataFormats/Math/interface/GeantUnits.h"
-#include "Geometry/MuonCommonData/plugins/DDGEMAngular.h"
+#include "DetectorDescription/Core/interface/DDAlgorithmFactory.h"
+#include "DetectorDescription/Core/interface/DDTypes.h"
+#include "DetectorDescription/Core/interface/DDAlgorithm.h"
+
+class DDGEMAngular : public DDAlgorithm {
+public:
+  //Constructor and Destructor
+  DDGEMAngular();
+  ~DDGEMAngular() override;
+
+  void initialize(const DDNumericArguments& nArgs,
+                  const DDVectorArguments& vArgs,
+                  const DDMapArguments& mArgs,
+                  const DDStringArguments& sArgs,
+                  const DDStringVectorArguments& vsArgs) override;
+
+  void execute(DDCompactView& cpv) override;
+
+private:
+  double startAngle;  //Start angle
+  double stepAngle;   //Step  angle
+  int invert;         //inverted or forward
+  double rPos;        //Radial position of center
+  double zoffset;     //Offset in z
+  int n;              //Mumber of copies
+  int startCopyNo;    //Start copy Number
+  int incrCopyNo;     //Increment copy Number
+
+  std::string rotns;        //Namespace for rotation matrix
+  std::string idNameSpace;  //Namespace of this and ALL sub-parts
+  std::string childName;    //Children name
+};
 
 using namespace geant_units::operators;
 
@@ -97,3 +131,5 @@ void DDGEMAngular::execute(DDCompactView& cpv) {
     copyNo += incrCopyNo;
   }
 }
+
+DEFINE_EDM_PLUGIN(DDAlgorithmFactory, DDGEMAngular, "muon:DDGEMAngular");
